Move align_contigs into separate file
